+++ docs/Changelog.md
@@ -5,6 +5,7 @@
 - Mapsforge themes compatibility [#100](https://github.com/mapsforge/vtm/issues/100)
 - Render themes: line symbol [#124](https://github.com/mapsforge/vtm/issues/124)
 - Render themes: stroke dash array [#131](https://github.com/mapsforge/vtm/issues/131)
+- POI Search example [#394](https://github.com/mapsforge/vtm/issues/394)
 - Mapsforge fix artifacts for zoom > 17 [#231](https://github.com/mapsforge/vtm/issues/231)
 - vtm-theme-comparator module [#387](https://github.com/mapsforge/vtm/issues/387)
 - Many other minor improvements and bug fixes